Add the main marvin files
Marvin.py is the primary wrapper file. robot.py is my scratch file with some examples that are currently commented out. That's basically my working directory. MarvinJunior.py is going to be similar to marvin with the hopes being the user can see the code in action before pushing to the robot using the turtle python package. This is all very much a work in progress.